exm feed add

+++ exm/modules/exmFeed/containers/feed/List.tsx
@@ -1,12 +1,12 @@
-import { mutations, queries } from "../../graphql";
-import { useMutation, useQuery } from "@apollo/client";
+import { mutations, queries } from '../../graphql';
+import { useMutation, useQuery } from '@apollo/client';
 
-import Alert from "../../../utils/Alert";
-import List from "../../components/feed/List";
-import React from "react";
-import Spinner from "../../../common/Spinner";
-import { confirm } from "../../../utils";
-import gql from "graphql-tag";
+import Alert from '../../../utils/Alert';
+import List from '../../components/feed/List';
+import WelcomeList from '../../components/feed/WelcomeList';
+import React from 'react';
+import { confirm } from '../../../utils';
+import gql from 'graphql-tag';
 
 type Props = {
   queryParams: any;
@@ -14,29 +14,54 @@ type Props = {
 };
 
 export default function ListContainer(props: Props) {
-  const { queryParams, contentType } = props;
-
-  const limit =
-    queryParams && queryParams.limit ? parseInt(queryParams.limit, 10) : 20;
+  const { contentType } = props;
 
   const feedResponse = useQuery(gql(queries.feed), {
     variables: {
-      limit,
-      contentTypes: [contentType || "post"],
-    },
+      contentTypes: [contentType || 'post']
+    }
   });
 
   const [deleteMutation] = useMutation(gql(mutations.deleteFeed));
   const [pinMutation] = useMutation(gql(mutations.pinFeed));
 
   if (feedResponse.loading) {
-    return <Spinner objective={true} />;
+    return null;
   }
+
+  const loadMore = () => {
+    const feedLength = feedResponse.data.exmFeed.list.length || 0;
+
+    feedResponse.fetchMore({
+      variables: {
+        skip: feedLength
+      },
+      updateQuery(prev, { fetchMoreResult }) {
+        if (!fetchMoreResult) {
+          return prev;
+        }
+
+        const fetchedExmFeed = fetchMoreResult.exmFeed.list || [];
+
+        const prevExmFeed = prev.exmFeed.list || [];
+
+        if (fetchedExmFeed) {
+          return {
+            ...prev,
+            exmFeed: {
+              ...prev.exmFeed,
+              list: [...prevExmFeed, ...fetchedExmFeed]
+            }
+          };
+        }
+      }
+    });
+  };
 
   const pinItem = (_id: string) => {
     pinMutation({ variables: { _id } })
       .then(() => {
-        Alert.success("Success!");
+        Alert.success('Success!');
 
         feedResponse.refetch();
       })
@@ -49,7 +74,7 @@ export default function ListContainer(props: Props) {
     confirm().then(() => {
       deleteMutation({ variables: { _id } })
         .then(() => {
-          Alert.success("You successfully deleted.");
+          Alert.success('You successfully deleted.');
 
           feedResponse.refetch();
         })
@@ -59,15 +84,23 @@ export default function ListContainer(props: Props) {
     });
   };
 
-  const { list, totalCount } = feedResponse.data?.exmFeed || {};
+  const exmFeed = feedResponse.data?.exmFeed.list || [];
+  const totalCount = feedResponse.data?.exmFeed.totalCount || 0;
+
+  if (contentType === 'welcome') {
+    return (
+      <WelcomeList list={exmFeed} totalCount={totalCount} loadMore={loadMore} />
+    );
+  }
 
   return (
     <List
       deleteItem={deleteItem}
       pinItem={pinItem}
-      list={list}
+      list={exmFeed}
       totalCount={totalCount}
-      limit={limit}
+      contentType={contentType}
+      loadMore={loadMore}
     />
   );
 }
